# internal/handler/device_handler.go
package handler

import (
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/google/uuid"
	"github.com/kiefernetworks/shellvault-server/internal/middleware"
	"github.com/kiefernetworks/shellvault-server/internal/repository"
)

type DeviceHandler struct {
	deviceRepo repository.DeviceRepository
	tokenRepo  repository.TokenRepository
}

func NewDeviceHandler(deviceRepo repository.DeviceRepository, tokenRepo repository.TokenRepository) *DeviceHandler {
	return &DeviceHandler{
		deviceRepo: deviceRepo,
		tokenRepo:  tokenRepo,
	}
}

func (h *DeviceHandler) ListDevices(w http.ResponseWriter, r *http.Request) {
	userID, ok := middleware.GetUserID(r.Context())
	if !ok {
		respondError(w, http.StatusUnauthorized, "unauthorized")
		return
	}

	devices, err := h.deviceRepo.GetByUserID(r.Context(), userID)
	if err != nil {
		respondError(w, http.StatusInternalServerError, "failed to list devices")
		return
	}

	if devices == nil {
		respondJSON(w, http.StatusOK, []interface{}{})
		return
	}

	respondJSON(w, http.StatusOK, devices)
}

func (h *DeviceHandler) DeleteDevice(w http.ResponseWriter, r *http.Request) {
	userID, ok := middleware.GetUserID(r.Context())
	if !ok {
		respondError(w, http.StatusUnauthorized, "unauthorized")
		return
	}

	deviceID, err := uuid.Parse(chi.URLParam(r, "id"))
	if err != nil {
		respondError(w, http.StatusBadRequest, "invalid device id")
		return
	}

	if err := h.deviceRepo.Delete(r.Context(), deviceID, userID); err != nil {
		respondError(w, http.StatusInternalServerError, "failed to delete device")
		return
	}

	respondJSON(w, http.StatusOK, map[string]string{"status": "device removed"})
}